requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.2
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
python-multipart==0.0.6
pydantic==2.5.2
firebase-admin==6.2.0
aiohttp==3.9.1
python-dotenv==1.0.0
pytest==7.4.3
python-whois==0.8.0
dnspython==2.4.2
tld==0.13
urllib3==2.1.0 